Added authorization bearer to the header of the request to the server.

+++ src/components/MyRecipeDetails.jsx
@@ -67,10 +67,15 @@ function MyRecipeDetails() {
 
     const handleDeleteRecipe = async () => {
         try {
+
+            const token = localStorage.getItem('token'); // Retrieve the token from localStorage
+            if (!token) throw new Error('Token not found');
+
             const response = await fetch(`https://ga-p3-backend.onrender.com/recipes/delete/${params.id}`, {
                 method: 'DELETE',
                 headers: {
                     'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${token}` // Include the authorization header
                 },
             });
